pyproject.toml
[tool.poetry]
name = "AFastapi"
version = "0.1.0"
description = "Asdil's fastapi template"
authors = ["Asdil <jpl4job@126.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
python-dateutil = "^2.8.2"
loguru = "^0.6.0"
PyJWT = "^2.4.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
PyMySQL = "^1.0.2"
DBUtils = "^3.0.2"
py2neo = "^2021.2.3"
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
